Keyboard struct implementation finished

+++ Structs_Enums_Interfaces/Structs_Enums_Interfaces/Enums/Note.cs
@@ -1,4 +1,4 @@
-﻿public enum Note: short
+﻿public enum Note: byte
 {
     Cb = 0,
     C = 1,
